+++ sdk/go/worker/worker_test.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/coretexos/cap/sdk/go"
-	agentv1 "github.com/coretexos/cap/go/coretex/agent/v1"
+	agentv1 "github.com/coretexos/cap/v2/go/coretex/agent/v1"
 	"github.com/nats-io/nats.go"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/timestamppb"