+++ sdk/python/cap/client.py
@@ -1,16 +1,21 @@
-import nats
 from google.protobuf import timestamp_pb2
 from cap.pb.coretex.agent.v1 import buspacket_pb2
 from cryptography.hazmat.primitives.asymmetric import ec
 from cryptography.hazmat.primitives import hashes
-import hashlib
+from typing import Optional
 
 
 DEFAULT_PROTOCOL_VERSION = 1
 SUBJECT_SUBMIT = "sys.job.submit"
 
 
-async def submit_job(nc, job_request, trace_id: str, sender_id: str, private_key: ec.EllipticCurvePrivateKey):
+async def submit_job(
+    nc,
+    job_request,
+    trace_id: str,
+    sender_id: str,
+    private_key: Optional[ec.EllipticCurvePrivateKey] = None,
+):
     ts = timestamp_pb2.Timestamp()
     ts.GetCurrentTime()
     packet = buspacket_pb2.BusPacket()
@@ -20,9 +25,9 @@ async def submit_job(nc, job_request, trace_id: str, sender_id: str, private_key
     packet.protocol_version = DEFAULT_PROTOCOL_VERSION
     packet.job_request.CopyFrom(job_request)
 
-    unsigned_data = packet.SerializeToString()
-    digest = hashlib.sha256(unsigned_data).digest()
-    signature = private_key.sign(digest, ec.ECDSA(hashes.SHA256()))
-    packet.signature = signature
+    if private_key:
+        unsigned_data = packet.SerializeToString()
+        signature = private_key.sign(unsigned_data, ec.ECDSA(hashes.SHA256()))
+        packet.signature = signature
 
     await nc.publish(SUBJECT_SUBMIT, packet.SerializeToString())